Make letter contact details wider
Typical contact details for services who send letters don’t fit in the
space we’d allotted.

This commit increases the width from a third column to a two-fifths
column. This is based on some experimentation that Sonia has been doing.

It will also need moving vertically, but I’ll do that work separately
once we’ve decided on the best alignment.