Let `Field`s accept list values as well as string
Our users want to put lists of variable length into their messages. The
elegant way of doing this (I reckon) is for users to pass a `list` as
the value of the placeholder.

Currently `Field`s only accept a `str` as the value of a placeholder.

This commit:
- handles the case where other code supplies a `list` as the value of a
  placeholder
- outputs that `list` as a comma-separated list, eg ‘one, two and three’

We don’t can’t replace lists with Markdown everywhere, because not all
templates, or parts of templates (eg an email subject) support Markdown
or HTML. We could raise an error in this case, but providing a
human-friendly, plain text fallback feels like a nice thing to do.