Format placeholder value lists as Markdown lists
The real need for lists as values of placeholders is in emails and
letters. This is where the message may contain, for example:

- a list of properties that a recipient is interested in
- a list of documents that the recipient needs to supply

These would be too long for a comma-separated list, and a bulleted list
is more appropriate. So this commit adds an option to `Field`s to return
a Markdown-formatted bulleted list instead.

A list of one item should still be presented bulleted, because the
content around it probably wouldn’t make sense if we folded a one-item
list up.