Don’t use default Python CSV `dict` reader
We have a need to extract lists from rows of CSV files. This will mean
customising how we parse CSV files into `list`s of dictionaries. The
default implementation throws away some of the data that we need.

We can still use that CSV reader that turns a CSV file into a list of
lists (list of rows each containing a list of columns – a 2 dimensional
array).

This means, to start with, implementing our own way of turning a `list`
representation of a row into a dictionary. But rather than implement it
from scratch this code reuses the Python CSV implementation, with a few
tweaks to make it fit in our code, and some variable renaming for
readability and style consistency.

The original code, from which this is adapted, can be found here:
https://github.com/python/cpython/blob/1e2147b9d75a64df370a9393c2b5b9d170dc0afd/Lib/csv.py#L108-L128