Make lists from same-named columns of a CSV file
We need a way for users to provide us with a list in their CSV file.
We could do this with newlines, or commas, or some other delimiter, but
this is bad because:
- we’re inventing our own syntax
- these things already have a specific meaning in CSV files
- there is already a natural delimiter in spreadsheets – the column

So we think that the elegant way of specifying this is:
- one column per list items
- all columns that are part of the same list are named the same

This is good because:
- no special syntax
- no limit on the number of items in a list
- doesn’t change how placeholders work in the template

In other words turn this spreadsheet:

phone number| list| list| list
---|---|---|---
07900900001| cat| rat| gnat

And this placeholder:
```
Here are some animals:

((list))
```

Into this `dict`:
```python
{'list', ['cat', 'rat', 'gnat']}
```

Which, in a template, gives this output:
> Here are some animals:
>
> * cat
> * rat
> * gnat

This commit does the part which turns the spreadsheet into the `dict`.

Unfortunately the code to do this is not so elegant, particularly the
dictionary mutation. But it is:
- tested
- still performant over a large CSV file