Dedupe column headers
If you provide a CSV file with multiple columns with the same name we
interpret that as a list – ie one placeholder with multiple values.

But the frontend preview of the file won’t reflect this; it will show
multiple columns each containing (the same) multiple values.

To stop this from happening, the column headers we return to the
frontend should be unique. We still need to keep the full, duplicated
version inside each instance of this class in order to populate lists
of values.