Output date and subject as part of letter body
DVLA have defined us separate boxes in the template for the date and
subject.

We have moved the date from the right to left, so it doesn’t need a
separate box. The subject never needed a separate box; it can just be a
`<h1>` in the content box.

This means that the same box can now hold:
- date
- subject
- content

So this commit:
- passes them all in via the `NOTIFICATION_BODY` field in the file
- adds formatting markup for the subject to make it big and bold

IMPORTANT – changes that DVLA will need to make as a result of this:

- delete the subject box
- delete the date box
- make sure they have the right font styles set up for the `<h1>`
- make sure that the vertical position of the content box is 95mm from
  the top margin, or 105mm from the top edge of the paper (double check
  these figures with Sonia and against the examples on the wall)