Handle any possible type as placeholder value
One of our users was passing an integer as the value of their
placeholder. We didn’t anticipate this, but it did work.

When we changed to explicitly allow placeholders to be lists as well as
strings we inadvertently started ignoring anything that wasn’t a list
or string. This meant that HTML for marking up the placeholder started
showing up in people’s text messages 😳

This commit adds code to handle (as sensibly as possible) any valid JSON
type, which are:
- number (integer or floating point)
- string
- boolean
- array
- object
- null

This is based on the principle that we should be liberal in what we
accept, and, by having tests for all these types, predictable.