Make list output conform to what works for DVLA
This is the format that we have tested and know gives us the visually
correct thing when we send it to DVLA:

```
Here is a list of bullets:  <cr><op><bul><tab>bullet 1  <op><bul><tab>bullet 2  <op><bul><tab>bullet 3  <p><cr><cr>And here is a list of long bullets…
```

This commit changes the output for bulleted lists to match this.

It doesn’t change the output for numbered lists, which may still be
incorrect.