Ensure right no. of newlines between para and list
The output that seems to space things out properly in DVLA-generated
letters is one `<cr>` between a paragraph and a list. By default we
insert two `<cr>`s after a paragraph, and one before a list.

This commit does some post-processing of the output to fix this.

There’s no way to do this in Markdown itself, because Markdown looks at
each element individually, not in combination.

Should check that this is also the correct thing to do for headings by
testing it with DVLA.