Never send DVLA anything for TO_NAME_2
Problem: we were sending the first line of the address in the
`TO_NAME_2` field. This meant that they couldn’t do any PAF lookups with
it, because it wasn’t where they were expecting.

The first line of the address is the second line of what our users give
us. We need to give this to DVLA as the _third_ line of the address
output, which they call `TO_ADDRESS_LINE_1`.

```
Our users give us      We store it as       We give it to DVLA as
========================================================================
First name-------------First name-----------TO_NAME_1
123 fake street--------123 fake street---|  TO_NAME_2 (left blank)
                   ┌---SW1A 1AA-----┐    └--TO_ADDRESS_LINE_1
                   |                |       TO_ADDRESS_LINE_2
                   |                |       TO_ADDRESS_LINE_3
                   |                |       TO_ADDRESS_LINE_4
                   |                |       TO_ADDRESS_LINE_5
SW1A 1AA-----------┘                └-------TO_POST_CODE

```