Make address line 2 mandatory
DVLA require:
- a recipient name
- the first line of the physical address
- a postcode

They check the first line of the physical address and the postcode
against the PAF.

We we need to make providing these lines mandatory. In our parlance
these things are:
- `address line 1`
- `address line 2`
- `postcode`

This commit adds `address line 2` to the set of mandatory address lines
that we need in an API call/CSV file.