Make `ORG_ID` in configurable in DVLA file
The `ORG_ID` tells DVLA which logo to put in the letter.

Currently it’s hard-coded to `500` (HM Land Registry). In the future we
will want to print letters which have the logos of other organisations.

This commit makes `org_id` an argument of the relevant `Template`
subclass, which is the same way we do this for email branding.

Still defaults to `500` so that this isn’t a breaking change.