Make logo configurable in letter previews
Currently letter previews are hard coded to use the HM Government logo.
While we may want this to be the default, we will in the future want to
give services (or let services pick) their own branding.

This commit makes the filename of the logo an argument of the relevant
`Template` subclass. The intention is that whatever code is initialising
this class will look up the correct logo filename to use from somewhere
and pass it in.

Still defaults to use the HM Government logo, so that this isn’t a
breaking change.