Strip DVLA markup from raw template content
We don’t want users to be able to control the formatting of their
letters by inserting the kind of pseudo-HTML that we send to DVLA. They
would have to be pretty savvy to figure out how to do this, but it seems
worth guarding against.

This commit also does the same for the previews, so that they remain an
accurate representation of what the printed letters will look like.