Fix empty address placeholders becoming HTML
Situation
---
A CSV file has an optional address column (eg `address line 3`) but no
values for this column in (some of) the rows.

Problem
---
The outputted data for those rows will have
`<span class='placeholder'>…` in the address block, instead of empty
strings.

Why it’s happening
---

`{'x': None}.get('x', '')` returns `None`, _not_ `''`.

`Field`s treat `None` as missing personalisation, and `''` as a valid
replacement. So they were seeing missing personalisation when the intent
of the code was to say _replace this bit of the address with an empty
string because it’s optional_.

The fix
---
Converting `None` to `''` for all optional address lines.