Generate image URLs for all pages of a letter
This is in order to show all pages of a letter in the admin app.

Takes the count of pages, returns HTML `<img>` tags for each page in the
letter.

Doesn’t (yet) remove the link to the PDF version because it’s probably
still useful to have it for our testing.