Use correct name for font
When the font is installed in the Docker container, running
`fc-list | tail -n3` returns this:
```
...
/usr/share/fonts/type1/gsfonts/p052024l.pfb: URW Palladio L:style=Bold Italic
/usr/share/fonts/truetype/dejavu/DejaVuSerif.ttf: DejaVu Serif:style=Book
/usr/share/fonts/type1/gsfonts/n019003l.pfb: Nimbus Sans L:style=Regular
```

This means that our CSS no longer references the font by the name it’s
called. And it doesn’t seem to be able to guess.

So this commit changes the CSS to use the name used by the Debian
package.