Validate international phone numbers
This commit adds optional extra validation that allows international
phone numbers. We consider an international phone number to be anything
that starts with:

- any number of `0`s, and then a valid country prefix
- `+` and then a valid country prefix
- just a valid country prefix

As with UK numbers we ignore any brackets, dashes, spaces before
validating.

There is one exception, Russia, whose country prefix is `7`. A number
starting with `7` could be either a UK mobile number or a Russian phone
number. So we only consider a phone number to be Russian if the part
after the `7` is at least 10 digits.

Any number that starts with `7` and has fewer than 10 digits we treat as
a UK mobile number. UK mobile numbers have 9 digits after the first `7`.