Treat single leading 0s as UK phone numbers
Before we discarded any number of leading zeros. This commit adds a
special case for phone numbers that start with exactly one `0`.

This is because we reckon that someone giving us `020 7946 0991` is more
likely to have accidentally got a UK landline number in their data,
rather trying to send to Egypt (country prefix `20`).