Add function to return info about a phone number
The API needs to know three things about a phone number:
- whether it’s international
- what the prefix is
- how many billable unit to charge for a text sent to this number

This commit adds a function to return a tuple of that information.

This only subtlety is what to do with a number like `1664000000000`,
which matches USA (`1`) and Montserrat (`1664`). In this case we return
the most precise match, ie Montserrat.