Remove the link from the image of the letter
I think we’ve established that people are happy enough seeing the image
of the letter in the browser. If there is a need to get a PDF of it,
then we should make an obvious route to doing so. Clicking the image
isn’t an obvious route.

So this commit removes the link from the template that generates the
HTML `<img>` tag for the letter.

Will need some CSS work in the admin app to keep things looking the
same.