Remove references to PDF link
This commit removes or renames a bunch of things which talked about
linking to a PDF. This does mean a breaking change for the admin app,
but it also means that the code remains clear about what it’s doing.

It also changes the `image_url` argument to require the full URL,
including `.png`. This means that we are generating the full URL in the
admin app, which is better because we can use the Flask helpers.