Use external library to format phone numbers
This commit brings in the `phonenumbers` library, which is a port[1] of
Google’s library for dealing with international phone numbers[2].

The useful thing it gives us is locale-specific formatting of phone
numbers. So in the USA the idiomatic way to format a phone number is:

> +1 202 555-5555

Whereas France doesn’t use dashes, and uses chunks of two digits:

> +33 1 12 34 56 78

So this commit changes our code to format numbers in whatever way people
from that country are used to. So someone trying to match up a phone
number (either visually or in conversation) with something they see in
Notify should have an easier time.

1. https://github.com/daviddrysdale/python-phonenumbers
2. https://github.com/googlei18n/libphonenumber