Stop bleach trying to create well-formed HTML
Bleach parses the text and tries to create valid HTML before cleaning
it. This means that when you pass in something that sort of looks like
HTML but isn’t meant to be, you get very different output.

This commit hacks around that by pre-encoding some of the HTML-like
stuff so that it no longer looks HTML-like to bleach. Replacing the `<`
with an entity is something that bleach would be doing in any case.