Allow missing placeholders to show up ‘redacted’
We have a need to not show things like one time passwords or two factor
codes when we show the content of messages.

We could just show the placeholders, but this might make people think
that placeholders have accidentally been sent out.

What we’re really doing is redacting the content of the message. And the
visual language of redaction is a thick black marker pen. This commit
replicates that in unicode, so that people are less likely to be
confused about what went out.