Make line-height of paragraphs match DVLA
Here is the state of DVLA’s software:
- It does line spacing in mm between lines, not height of lines in
  points
- It can can only specify line spacing to a maximum precision of 0.1mm

We told DVLA to set the line spacing to 1.23mm (which is 16 − 12.5pt).
They set it to 1.2mm, which is as precise as they can go. This means a
difference of 0.03mm, or 0.1pt. Doesn’t sound like a lot, but over a few
the course of a few lines it’s noticeable different. And over the course
of a page it could be the difference between a letter fitting on one
page or two.

This commit make the letter preview match the line spacing of what DVLA
are producing, reducing it by 0.03mm, or 0.1pt (0.03 ÷ 25.4 × 72).