Reduce space between paragraphs to match DVLA
DVLA’s software seems to set the space between paragraphs to equal the
size of the font, not the line height as we have it in our
preview/designs.

Let’s capitulate for now and try to get our preview matching the DVLA
thing, so that our page counts match up.