Remove whitespace before commas in things we send
I’ve seen a couple of times (and we’ve done it in emails our service has
sent) where a space sneaks in before a comma. This tends to happen with
placeholders, eg:

> Hello ((name)) ,

…which results in:

> Hello Chris ,

…which doesn’t look great.

Now we could do some postprocessing of the template content to remove
whitespace between closing placeholder tags and commas, ie look for
anything like `)) ,` and strip it before saving the template.

I think it’s better to do it at the time of rendering because:

- this covers templates already in the system
- I can’t think of a single case where we’d _ever_ want a space before a
  comma (butt in here if you can)

So this commit does the replacement in all the places where it might be
useful, for all types of messages.

The tests are fairly verbose here, but this stuff is quite complex now,
so more tests can’t hurt I don’t think.