Fix bug with formatting recipients for whitelist
When we compare recipients to those in the whitelist we have a special
function to normalise them to make sure we’re comparing like with like.

I wasn’t tested at all, and was causing bugs elsewhere in the code
because it didn’t always return strings.

This commit fixes that bug and adds a unit test.

I have an integration test to go in admin which targets the specific
bug.