Remove whitespace before full stops and commas
We already remove whitespace before commas. We should do the same for
full stops. It’s the kind of nice thing that the iPhone keyboard does
for you, and looks unprofessional when left uncorrected.

Can’t think of a use case where we wouldn’t want this (apart from in
URLs, but they don’t contain spaces anyway).