Replace dumb quotes with smart quotes
> “Smart quotes,” the correct quotation marks and apostrophes, are curly
> or sloped. "Dumb quotes," or straight quotes, are a vestigial
> constraint from typewriters when using one key for two different marks
> helped save space on a keyboard. Unfortunately, many improper marks
> make their way onto websites because of dumb defaults in applications
> and CMSs.

Most people don’t know this though. And most people writing email or
letter templates don’t. So let’s do the hard work for them.

This commit brings in a module called Smartypants[1] which replaces dumb
quotes with smart ones in a sensible way – ie it takes the context into
consideration. It’s not quite as smart as Medium’s implementation[2],
but since theirs is closed source we can’t use it 😔.

Smartypants can do quite a few things, but this restricts it to only
replacing quotes. And configures it to replace them with the unicode
equivalent, rather than encoded entities (because I doubt DVLA’s weird
markup understands HTML entities). That’s what `Attr.s | Attr.u` means –
worth calling out because it’s such a weird API.

We don’t want to do this for text messages because we’d only go and
downgrade the characters again.

1. http://pythonhosted.org/smartypants/
2. https://medium.design/quotation-marks-c8993b54417c