Add methods put/remove content in empty newlines
By default, Markdown collapses multiple newlines into one (or, more
precisely into two paragraphs). In other words, it ignores lines that
are empty, and assumes they are just there to demarcate two blocks of
text.

Because this is a fundamental tenet of Markdown, it’s a behaviour that’s
hard to override in the parser (which is what we’ve done for other
custom features). Instead we have to hack around it.

Because Markdown ignores empty lines, we can make it pay attention to
those lines by making them not empty. So this commit adds two functions:

1. One to fill sequences of empty lines longer than one with some
   special characters that make them non-empty.

2. One to remove the special characters after the Markdown parsing is
   done.

The special characters were chosen to be something that no-one would
type in an actual message.