Allow multiple newlines in letter templates
Normally, the default Markdown behaviour of collapsing multiple newlines
is desirable, because we want consistency in the amount/size of
whitespace appearing in our emails and letters.

However, for letters, we need to give teams control over vertical space
so that they can push content onto the next page of a letter. This is
because they have a legislative obligation to do this. And while it’s a
hacky way of meeting that obligation (compared to, say, a page break)
it’s the best we can offer for now.