Make newline substitution carriage–return-aware
We detect and replace newlines in letter templates to override
Markdown’s default newline handling.

This was working fine in theory, but in the real world newline
characters sometimes come combined with carriage return characters (or,
on Mac OS 9, just carriage return characters). For example, this plain
text stored in a variable called `template`:
```
one
two
```

Could come out as:
```python
repr(template)
>>> 'one\r\ntwo'
```

…or:
```python
repr(template)
>>> 'one\ntwo'
```

Our substitution doesn’t work in the former case because it’s only
expecting `\n` characters. And it turns out that the former case is the
reality of what the API returns (either because of browsers/Flask/Mac
OS; I didn’t dig that deep into it).

So this commit normalises the string to only have `\n` characters
before doing any substitution on it.