Tweak line spacing for uniformity with DVLA
The number of lines fitting on a page still wasn’t quite right. We have
sample content we’re using to test with DVLA – this commit tweaks the
line spacing until the DVLA output and our preview both fit the same
number of lines on a page.

We can do this now that all other inconsistencies are eliminated.