Make delivery estimates default to 4pm
The delivery estimate should be able to give you a reasonable chance of
guessing whether a letter has been delivered or not.

According to Royal Mail:

> We aim to deliver to addresses in urban areas by 3pm every working day
> and addresses in rural areas by 4pm every working day

– http://www.royalmail.com/sites/default/files/docs/pdf/106831_col_del_pdf_02.pdf..unpublished

So if you wait until after 4pm you can be reasonably sure that a letter
has been delivered.

Only changing this because the API _has_ to return an hour (to be a
valid timestamp) so it might as well return the most sensible hour I can
think of.