Ensure datetimes returned are UTC
We deal with UTC everywhere, up to the point where we present a time
back to the user on the frontend. So all code that is not specifically
to do with presenting a date on the frontend should take dates assuming
they’re in UTC (done) and return them as UTC. This commit does the
latter.