Fix over-greedy whitespace replacement in address
We run a formatter called `remove_whitespace_before_punctuation` on our
address blocks, which is supposed to fix typos like ‘Ms . Example’ or
‘Cheltenham    , Gloucestershire’.

This had the unfortunate side effect of looking not just at tabs and
spaces, but also newlines. Which meant that if an address line started
with a punctuation character, it would get merged with the preceding
address line.

In other words this:
```
Firstname Lastname
, 123 Example Street
SW1A 1AA
```

…would get transformed to:
```
Firstname Lastname, 123 Example Street
SW1A 1AA
```

This means that we have a possible variation in the number of address
lines. Other parts of our codebase were expecting us to always return
8 lines, and so were raising an exception.

This commit changes our code to look only for tabs and spaces, not other
whitespace characters (ie any kind of newline, carraige return, etc.)