Hide barcodes and MDI from printing
The barcodes we show in the preview will be added by DVLA in the printed
version of the template. So we shouldn’t be including them ourselves in
the printed PDF.

This commit adds some extra CSS, to the print template only, that hides
these things from the print template.