Allow spreadsheet rows to be looked up by index
In the future we want to be able to show a preview of any message in
your file. This means being able to get the value of any row in the
spreadsheet.

Currently the admin app looks at the first item in the `.rows`
property of `RecipientCSV`[1].

This commit changes the class to allow access to any row in the file. It
does this by implementing the `__getitem__` method, which is the magic
method that gets called when evaluating `instance[i]`.