Pin html5lib to last working version
html5lib has released version 1.0.1 today. html5lib is a dependency of
a package we depend on (bleach).

This seems to have some breaking changes. As an indirect dependency, we
don’t control which version of html5lib we’re using (as a matter of
practice). In this case though, we do need to control it, so that our
apps get the latest working version.