Add support for horizontal rules in emails
We’re finding that the GOV.UK email alerts are quite complex in terms of
hierarchy.

We think that a horizontal rule (with some spacing around it) is an
effective way to break up the content a bit, thereby making it clearer
what the structure of the content is.

This will be an undocumented feature to start with, though we think it
will be useful to more than just the GOV.UK email team.