Enable markdown link syntax for emails
We think there is a problem with showing full, long URLs in emails,
when an email has lots of links in it. These long links interrupt the
structure of the message and makes it:
- harder to read
- harder to see which link to click

For emails with just one link (typically a call to action) we think that
showing the full URL is still preferable.

This commit switches on the ability for teams to use Markdown syntax for
links with text labels, ie `[Link text](https://example.com)`.

This feature is for emails only. We will not documented it to start
with, and it should not be relied upon (ie it could go away).