Be stricter about length of phone numbers
We’ve been trying to send messages to some 5 digit phone numbers. This
is causing our provider to return errors. We should be catching these
before they get to the provider.

Stackoverflow seems to suggest[1] that the shortest phone numbers in use
in the world are 4 digits long (Niue). However we don’t support sending
messages to this country (not in the list of countries we know how to
bill for).

So the shortest real phone number we can send to is 5 digits long.
Countries with 5 digit phone numbers include the Solomon Islands and the
Cook Islands. They both have 3 digit international dialling codes, for
a total of 8 digits:
```
682 12345
 3 +  5  = 8
```

This commit updates our validation to reject anything shorter than 8
characters.