Truncate wide addresses in letters
When a line of text was too wide for the address block it was being
cut off. This is the behaviour we want.

However, we break up the address lines with `<br>` elements. These
elements get pushed onto the next line when the text is wider than its
container.

This caused extra linebreaks in our addresses, which appeared after
lines that were too wide for the address block.

This is fixed by treating each line of the address as its own
block-level element. List items seemed like an appropriate choice.