Downsize logos that are too wide for letters
Some logos are really wide. We were displaying them at full height, but
cropping off the right-hand side of them. This looks _bad_.

This commit changes the CSS to always show the full logo, and scale it
down if it’s too wide to fit in the containing box. The value of 70mm
we have for the containing box seems to look pretty good.

This doesn’t affect the display of existing, narrow logos – they still
take up the full 18mm height of the logo box.