Remove code that represents letters as DVLA markup
Now that we’re sending letters as PDF none of this is needed.

This only removes the code that won’t have an impact on PDF letters. For
example we strip DVLA markup from people’s templates in both kinds of
letter; if we stopped doing this it would change any letters that
include `<cr>`, for example.