Handle a CSV file with duplicate recipient columns
When multiple columns have the same name we append their values to a
list. This lets people make variable length lists.

However if the phone number, email, or address validation receives a
list of values, it doesn’t know what to do with it. And we can’t
gracefully handle a list of phone numbers because which one would we
send to?

Clearly in this case something is wrong with the file, so we should be
able to tell the user that. This commit:
- stops an exception being thrown in this case
- returns data from the instance which will help us construct a useful
  error message in the admin app