Add Markdown renderer for plain text emails
At the moment we just give users raw markdown in their plain text
emails. Markdown is _mostly_ plain text, but:

- there are enough parts of the syntax that will look weird to someone
  who is just expecting text (eg links)
- we can do a better job of breaking up the emails visually with some
  extra spacing and use of

As recommended here[1] we are not hard wrapping lines because most
clients these days will soft-wrap, and hard wrapping causes the text to
reflow really badly.

The only exception to this is the visual lines under headings and those
created to simulate a HTML `<hrule>`. These we are truncating at 65
characters, which is about as wide as they can go without them being
wrapped in most clients.

1. https://www.campaignmonitor.com/blog/email-marketing/2012/08/removing-the-60-character-line-limit-on-plain-text-email/