Make logos in branding banners 54px high
This matches the other kind of off-brand logos we support, those on a
white background.

32px matches the GOV.UK Crown, but is too small for most council logos
to be legible.

All the logos we have uploaded to live at the moment are 108px (54px ×
2, for retina screens) or bigger.

Fixed height on the table plus capital ‘M’ in margin seems to be the
best combination of hacks to get it looking right in Outlook.

Email on Acid test results:
https://www.emailonacid.com/app/acidtest/ji1JfgtUeqqB3cbS5DuHYor0fX6z85FldXKdJUp2GjXXb/list