Don’t truncate sequences of multiple asterisks
We have a user who want to put a sequence of asterisks in their email in
order to show a credit card number with all but the last four digits
obscured.

At the moment this doesn’t work, because Markdown interprets asterisks
as emphasis.

So `******` gets interpreted as `Start emphasis` `**` `End emphasis`.
But since we disregard emphasis what the user gets back in their email
template is `**`.

This commit makes the emphasis-formatting pass through any asterisks,
rather than strip them out. Which should result in the behaviour the
user is expecting when typing a long series of asterisks.

It’s also possible to do emphasis with underscores, therefore someone
typing `_foo_` will get back `*foo*`, and `______` will become `**__**`.
But I think that these cases are unlikely enough that the benefit of
passing through the asterisks outweighs this potential weirdness.