feat(sms): switch to direct Telstra client
Keep the Identity client in place and fallback to it if the Telstra
client fails.

Note: the env vars changed in order to make it clear to which SMS client
they apply.

Note: the Telstra client doesn't account for provisioning the number -
i.e. it assumes that it has been provisioned separately.