feat(builds): use zero-downtime-deploys for API
using zero-downtime-deploy means that everything for the app must be
contained in a manifest. because we don't want to store sensitive data
in those comitted files, we need to use user-provided-services in
cloudfoundry.

These allow us to create a custom service essentially just for its
environment variables, and then bind it to an application. To expose
those vcap variables into the running program we've used a wrapper
called vcapenvwrapper that lives at https://github.com/govau/vcapenv