fix(tests): make test-notifications
Use AU mobile numbers.

Add newlines to template content.

When a template (of class PlainTextEmailTemplate) is converted to a
string, it gets a newline added to it. It might be better to assert the
newline wherever necessary rather than defining it as part of the test
fixture, but we'll see how it goes.