feat(tests): use local postgres by default
And allow fallback to docker.

We also require database explicity in some tests that just happened to
work due to a leftover database state between tests on a non-worker
database instance. Now all tests that require database state must state
it explicity or pull a dependency that uses it. This forces them to use
a properly set-up and torn-down worker database instance; and lets us
avoid running migrate on the root instance.