fix(tests): make test-delivery
To fix these tests, it was necessary to update the twilio and telstra
provider details in the database. Twilio now has a priority of 10
(highest), Telstra has a priority of 15, MMG and Firetext are both
disabled and now have a priority of 20. Twilio is also updated to
support international sending.

It is still worth removing MMG and Firetext-related code and records
since they are no longer necessary and unsupported.