fix(tests): don't use created_by_id in migration
Setting the user ID in these migrations causes tests to fail because the
users table cannot be dropped. It cannot be dropped because there is a
reference to the user in the provider_details table. This causes a
cascade effect of trying to fix the tests in relation to notify_db,
notify_db_session and restore_provider_details. I'm not willing to fix
the root cause at this stage, so the easiest solution is to just remove
the created_by_id reference in the migrations. The only real downside of
not having the created_by_id column set in the records is that we don't
know that it was actually the system user that made the changes.