fix(manifest): buildpack->buildpacks
CF now prefers using buildpacks instead of buildpack in the manifest
file.

Also formatted the YAML file using prettier.

The error when pushing was:

Deprecation warning: Use of 'buildpack' attribute in manifest is
deprecated in favor of 'buildpacks'. Please see
http://docs.cloudfoundry.org/devguide/deploy-apps/manifest.html#deprecated
for alternatives and other app manifest deprecations. This feature
will be removed in the future.