feat(sentry): add to admin, api and celery
api and celery are using the same sentry project.

Because of the difficulty of setting different env vars between dev,
staging, prod, CI, admin backend and admin frontend, we're just using
sentry environment: "prod" and "dev". "prod" is for production, "dev" is
for everything else including dev, feature branches and staging.

Add JS sourcemaps to frontend. Unclear if they work properly, but it's a
start.