fix(telstra): remove it from active providers
When sending a message, if it throws an exception (e.g. Twilio throws an
exception because it tries to send to an invalid number), Notify tries
to switch providers. This causes our provider to switch from Twili to
Telstra. Disabling Telstra for now until we can either support Telstra
properly or work out how to do this better.