report: exclude services created by platform admin
When a platform admin creates a service, we usually want to exclude it
from reports because it is often a test or dummy service. We can exclude
it from reports using the count_as_live field. This is set to false for
new services that are created by platform admins. Services of this
nature were already being excluded in the live services report. They are
now excluded from the trial services report too.

It's worth mentioning that count_as_live doesn't make that much sense in
the context of a trial service (since it's not live). We leave the field
named as such for now to keep it aligned with the original GDS codebase.
At the time of writing it appears that a better name is possibly
something like 'include in reports'.